origin: add build time and version info
Set ldflags for version such that apiserver and controller-manager version
flags report correct values instead of "UNKNOWN".

The method for looking up the catalog version is by tracing the commit history
merges and looking for the latest catalog rebase. This is done because there
isn't really a good version number to extract from the source (and upstream
builds are done using upstream tags).